Simplification Dijkstra no valid path part
Changed the used constructor to match the one used in the Bellamn-Ford algo
About the Junit failure, the path isn't valid, so the path is null. That is actually normal. The issue comes from the test, that was designed to compare values, not check the status of the solution. As this test isn't relevant, the solution is to change the points so that it doesn't fail anymore.
Origin/Destination changes to come in the next commit.
This O/D couple can be saved for a status Junit test that can be implemented, even if simply taking an island map and try to check from 2 separate island would work just as well.
Other thing to do : change the test map. This one is fine for no oracle tests, but BF is quite slow already, and taking 25 minutes when not all the tests are written is already too long for what we are trying to test.